schema: widen id to BigInteger on the 4 tables that grow at scale
registrations/maker_category_totals/fuel_category_totals/maker_fuel_totals
all delete-then-insert on every scrape/backfill rewrite, burning a fresh
autoincrement id on every rewrite, not just on net row growth -- confirmed
live, registrations' max(id) was already 80.6M against ~21.6M live rows.
Integer's ~2.1B ceiling is decades away at current growth rates, but the
model type change is free to make now and increasingly costly to make
later, so made it now rather than waiting for it to become urgent.

Model change alone only affects fresh installs (Base.metadata.create_all
has nothing to migrate for a new table); added ensure_bigint_id to actually
widen the column on this and any other already-deployed database, with an
idempotent no-op fast path once already bigint. Ran for real against the
dev DB: registrations took 22.7 minutes (a full table rewrite over 21.6M
rows, comparable in kind to the already-documented ~6min first-run index
build cost for this same table), the 3 smaller crosstab tables 27s-102s
each. Confirmed idempotent: a second pass across all 4 tables took 0.01s.

Left the small reference/lookup tables (states, rtos, users, ...) as
Integer -- no realistic path to billions of rows for any of them.

+++ backend/app/models/models.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, Float, Date, DateTime, Boolean, Index, text
+from sqlalchemy import BigInteger, Column, Integer, String, Float, Date, DateTime, Boolean, Index, text
 from sqlalchemy.sql import func
 from app.core.database import Base
 
@@ -44,7 +44,13 @@ class RTODistrict(Base):
 class Registration(Base):
     __tablename__ = "registrations"
 
-    id = Column(Integer, primary_key=True, autoincrement=True)
+    # BigInteger, not Integer: this table's own delete-then-insert write
+    # pattern burns a fresh id on every scrape/backfill rewrite, not just on
+    # net row growth -- max(id) was already 80.6M against ~21.6M live rows
+    # when checked. Free to set correctly now; a live ALTER on this table
+    # once it's actually approaching Integer's ~2.1B ceiling would cost far
+    # more than this one-line change does today.
+    id = Column(BigInteger, primary_key=True, autoincrement=True)
     state_code = Column(String(5), nullable=False, index=True)
     state_name = Column(String(100), nullable=False)
     rto_code = Column(String(10), nullable=True)
@@ -146,7 +152,9 @@ class MakerCategoryTotal(Base):
     dimension."""
     __tablename__ = "maker_category_totals"
 
-    id = Column(Integer, primary_key=True, autoincrement=True)
+    # BigInteger -- see Registration.id's comment; same delete-then-insert
+    # write pattern burns ids on every rewrite, not just net growth.
+    id = Column(BigInteger, primary_key=True, autoincrement=True)
     state_code = Column(String(5), nullable=False, index=True)
     state_name = Column(String(100), nullable=False)
     rto_code = Column(String(10), nullable=True)
@@ -193,7 +201,8 @@ class FuelCategoryTotal(Base):
     never both)."""
     __tablename__ = "fuel_category_totals"
 
-    id = Column(Integer, primary_key=True, autoincrement=True)
+    # BigInteger -- see Registration.id's comment.
+    id = Column(BigInteger, primary_key=True, autoincrement=True)
     state_code = Column(String(5), nullable=False, index=True)
     state_name = Column(String(100), nullable=False)
     rto_code = Column(String(10), nullable=True)
@@ -232,7 +241,8 @@ class MakerFuelTotal(Base):
     ICE/Hybrid/EV at query time via fuel_group(), same as FuelCategoryTotal."""
     __tablename__ = "maker_fuel_totals"
 
-    id = Column(Integer, primary_key=True, autoincrement=True)
+    # BigInteger -- see Registration.id's comment.
+    id = Column(BigInteger, primary_key=True, autoincrement=True)
     state_code = Column(String(5), nullable=False, index=True)
     state_name = Column(String(100), nullable=False)
     rto_code = Column(String(10), nullable=True)
